+++ history.md
@@ -1,4 +1,4 @@
-## 1.9.0 - TBD
+## 1.9.0 - 2015-12-16
 * Fixed a bug where the Response object of okhttp2 client only included the last header with a given name
 * Response::getHeaders now returns an unmodifiable collection instead of an Iterable.